+++ packages/@aws-cdk/aws-ecs/test/ec2/integ.lb-awsvpc-nw.ts
@@ -40,6 +40,6 @@ listener.addTargets('ECS', {
   targets: [service]
 });
 
-new cdk.Output(stack, 'LoadBalancerDNS', { value: lb.dnsName, });
+new cdk.CfnOutput(stack, 'LoadBalancerDNS', { value: lb.dnsName, });
 
 app.run();